.claude/worktrees/keen-shaw-651613/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
openai==1.30.1
anthropic==0.28.0
httpx==0.27.0
requests==2.33.1
pillow==10.3.0
aiosqlite==0.20.0
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
rembg==2.0.56
onnxruntime==1.18.0
